+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
-## v0.26.1
+## v0.27.0
 
 - remove invalid property from `Result`
 - add catalog class to dcat ontology
 - fix documentation
 - `Thing`s now have a validation method `validate()`, which uses SHACL shapes to validate the instance against the
   ontology schema
+- updated package setup (removed requirements.txt, following standard practices for python packages and use
+  pyproject.toml)
 
 ## v0.26.0